# DMS_api_gateway/docs/swagger.yaml
basePath: /
definitions:
...
  entity.UpdateNameUserInput:
    properties:
      name:
        type: string
    required:
    - name
    type: object
...
  entity.UpdatePasswordUserInput:
    properties:
      password:
        type: string
    required:
    - password
    type: object
...
  handler.UserResponse:
    properties:
      email:
        type: string
      username:
        type: string
    type: object
  handler.errorResponse:
    properties:
      message:
        type: string
    type: object
...
  handler.statusResponse:
    properties:
      status:
        type: string
    type: object
host: localhost:8000
info:
  contact: {}
  description: API Server for Document Management Service
  title: DMS API
  version: "1.0"
paths:
...
  /user:
    delete:
      consumes:
      - application/json
      description: deletes user account
      operationId: delete-user-account
      produces:
      - application/json
      responses:
        "200":
          description: deleted
          schema:
            $ref: '#/definitions/handler.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete User
      tags:
      - user
    get:
      consumes:
      - application/json
      description: gives information about the user
      operationId: get-user-info
      produces:
      - application/json
      responses:
        "200":
          description: user
          schema:
            $ref: '#/definitions/handler.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get User
      tags:
      - user
  /user/name:
    put:
      consumes:
      - application/json
      description: updates username
      operationId: update-name
      parameters:
      - description: update password
        in: body
        name: updateData
        required: true
        schema:
          $ref: '#/definitions/entity.UpdateNameUserInput'
      produces:
      - application/json
      responses:
        "200":
          description: updated name
          schema:
            $ref: '#/definitions/handler.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update Name
      tags:
      - user
  /user/password:
    put:
      consumes:
      - application/json
      description: updates user password
      operationId: update-password
      parameters:
      - description: update password
        in: body
        name: updateData
        required: true
        schema:
          $ref: '#/definitions/entity.UpdatePasswordUserInput'
      produces:
      - application/json
      responses:
        "200":
          description: updated password
          schema:
            $ref: '#/definitions/handler.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update Password
      tags:
      - user
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"